More type safety for Fields.
* Assert that the Field type is compatible in get().

* Support a static_cast-like transformation of field value, with
precision check.

+++ dbms/src/Interpreters/MetricLog.cpp
@@ -49,7 +49,7 @@ void MetricLogElement::appendToBlock(Block & block) const
         columns[column_idx++]->insert(profile_events[i]);
 
     for (size_t i = 0, end = CurrentMetrics::end(); i < end; ++i)
-        columns[column_idx++]->insert(current_metrics[i]);
+        columns[column_idx++]->insert(static_cast<Int64>(current_metrics[i]));
 }
